do not scan web extensions in desktop
